+++ internal/kernel/pkg/app/config/types.go
@@ -6,6 +6,7 @@ import (
 	"github.com/number571/go-peer/pkg/crypto/scheme/layer1"
 	"github.com/number571/go-peer/pkg/crypto/scheme/layer2"
 	logger "github.com/number571/hidden-lake/internal/utils/logger/std"
+	"github.com/number571/hidden-lake/pkg/api/kernel/client/scheme"
 )
 
 type IWrapper interface {
@@ -20,6 +21,7 @@ type IEditor interface {
 type IConfigSettings interface {
 	layer1.ISettings
 
+	GetCryptoSchemeType() scheme.ISchemeType
 	GetMessageSizeBytes() uint64
 	GetFetchTimeout() time.Duration
 	GetQueuePeriod() time.Duration