+++ internal/services/filesharer/internal/handler/incoming/load_test.go
@@ -14,6 +14,7 @@ import (
 	hlk_settings "github.com/number571/hidden-lake/internal/kernel/pkg/settings"
 	std_logger "github.com/number571/hidden-lake/internal/utils/logger/std"
 	hlk_client "github.com/number571/hidden-lake/pkg/api/kernel/client"
+	"github.com/number571/hidden-lake/pkg/api/kernel/client/scheme"
 	hls_config "github.com/number571/hidden-lake/pkg/api/kernel/config"
 	"github.com/number571/hidden-lake/pkg/network/request"
 	"github.com/number571/hidden-lake/pkg/network/response"
@@ -320,7 +321,7 @@ func (p *tsHLSClient) GetOnlines(context.Context) ([]string, error) {
 }
 func (p *tsHLSClient) DelOnline(context.Context, string) error { return nil }
 
-func (p *tsHLSClient) GetFriends(context.Context) (map[string]layer2.IParticipantKey, error) {
+func (p *tsHLSClient) GetFriends(context.Context, scheme.ISchemeType) (map[string]layer2.IParticipantKey, error) {
 	return map[string]layer2.IParticipantKey{
 		"abc": asymmetric.NewPrivKey().GetPubKey(),
 	}, nil