+++ internal/services/filesharer/internal/handler/stream/stream_test.go
@@ -14,6 +14,7 @@ import (
 	hls_filesharer_settings "github.com/number571/hidden-lake/internal/services/filesharer/pkg/settings"
 	"github.com/number571/hidden-lake/internal/utils/api"
 	hlk_client "github.com/number571/hidden-lake/pkg/api/kernel/client"
+	"github.com/number571/hidden-lake/pkg/api/kernel/client/scheme"
 	hls_config "github.com/number571/hidden-lake/pkg/api/kernel/config"
 	fileinfo "github.com/number571/hidden-lake/pkg/api/services/filesharer/client/dto"
 	"github.com/number571/hidden-lake/pkg/network/request"
@@ -231,7 +232,7 @@ func (p *tsHLSClient) GetSettings(context.Context) (hls_config.IConfigSettings, 
 func (p *tsHLSClient) GetOnlines(context.Context) ([]string, error) { return nil, nil }
 func (p *tsHLSClient) DelOnline(context.Context, string) error      { return nil }
 
-func (p *tsHLSClient) GetFriends(context.Context) (map[string]layer2.IParticipantKey, error) {
+func (p *tsHLSClient) GetFriends(context.Context, scheme.ISchemeType) (map[string]layer2.IParticipantKey, error) {
 	return nil, nil
 }