+++ internal/services/messenger/internal/handler/incoming/push_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/number571/hidden-lake/internal/utils/broker"
 	std_logger "github.com/number571/hidden-lake/internal/utils/logger/std"
 	hlk_client "github.com/number571/hidden-lake/pkg/api/kernel/client"
+	"github.com/number571/hidden-lake/pkg/api/kernel/client/scheme"
 	hls_config "github.com/number571/hidden-lake/pkg/api/kernel/config"
 	message_dto "github.com/number571/hidden-lake/pkg/api/services/messenger/client/dto"
 	"github.com/number571/hidden-lake/pkg/network/request"
@@ -168,7 +169,7 @@ func (p *tsHLSClient) GetOnlines(context.Context) ([]string, error) {
 }
 func (p *tsHLSClient) DelOnline(context.Context, string) error { return nil }
 
-func (p *tsHLSClient) GetFriends(context.Context) (map[string]layer2.IParticipantKey, error) {
+func (p *tsHLSClient) GetFriends(context.Context, scheme.ISchemeType) (map[string]layer2.IParticipantKey, error) {
 	return map[string]layer2.IParticipantKey{
 		"abc": p.fFriendPubKey,
 	}, nil