+++ internal/services/pinger/internal/handler/index_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/number571/go-peer/pkg/logger"
 	std_logger "github.com/number571/hidden-lake/internal/utils/logger/std"
 	hlk_client "github.com/number571/hidden-lake/pkg/api/kernel/client"
+	"github.com/number571/hidden-lake/pkg/api/kernel/client/scheme"
 	hlk_config "github.com/number571/hidden-lake/pkg/api/kernel/config"
 	hls_client "github.com/number571/hidden-lake/pkg/api/services/pinger/client"
 	"github.com/number571/hidden-lake/pkg/network/request"
@@ -101,7 +102,7 @@ func (p *tsHLKClient) GetOnlines(context.Context) ([]string, error) {
 }
 func (p *tsHLKClient) DelOnline(context.Context, string) error { return nil }
 
-func (p *tsHLKClient) GetFriends(context.Context) (map[string]layer2.IParticipantKey, error) {
+func (p *tsHLKClient) GetFriends(context.Context, scheme.ISchemeType) (map[string]layer2.IParticipantKey, error) {
 	return map[string]layer2.IParticipantKey{
 		"abc": asymmetric.NewPrivKey().GetPubKey(),
 	}, nil