+++ pkg/api/kernel/client/types.go
@@ -6,6 +6,7 @@ import (
 	"github.com/number571/go-peer/pkg/crypto/scheme/layer2"
 
 	friend "github.com/number571/hidden-lake/pkg/api/kernel/client/dto"
+	"github.com/number571/hidden-lake/pkg/api/kernel/client/scheme"
 	"github.com/number571/hidden-lake/pkg/api/kernel/config"
 	"github.com/number571/hidden-lake/pkg/network/request"
 	"github.com/number571/hidden-lake/pkg/network/response"
@@ -18,7 +19,7 @@ type IClient interface {
 	GetOnlines(context.Context) ([]string, error)
 	DelOnline(context.Context, string) error
 
-	GetFriends(context.Context) (map[string]layer2.IParticipantKey, error)
+	GetFriends(context.Context, scheme.ISchemeType) (map[string]layer2.IParticipantKey, error)
 	AddFriend(context.Context, string, layer2.IParticipantKey) error
 	DelFriend(context.Context, string) error
 
@@ -37,7 +38,7 @@ type IRequester interface {
 	GetOnlines(context.Context) ([]string, error)
 	DelOnline(context.Context, string) error
 
-	GetFriends(context.Context) (map[string]layer2.IParticipantKey, error)
+	GetFriends(context.Context, scheme.ISchemeType) (map[string]layer2.IParticipantKey, error)
 	AddFriend(context.Context, *friend.SFriend) error
 	DelFriend(context.Context, *friend.SFriend) error